% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{function1}
\alias{function1}
\title{Add two numbers}
\usage{
function1(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function adds two numbers together.
}
